Split drawing and shading functions into separate files
